FinanceVault/requirements.txt
pandas>=2.0.0
pdfplumber>=0.10.0
ofxparse>=0.21
tabulate>=0.9.0
click>=8.1.0
rich>=13.0.0
python-dateutil>=2.8.0
flask>=3.0.0
flask-cors>=4.0.0
